% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.R
\name{Batcheffect.Test}
\alias{Batcheffect.Test}
\title{Test for batch effect}
\usage{
Batcheffect.Test(n0, n1, n.ext, maf0, maf1, maf.ext, pop.prev, var.ratio = 1)
}
\arguments{
\item{n0}{A numeric. The sample size of cases in the study cohort}

\item{n1}{A numeric. The sample size of controls in the study cohort}

\item{n.ext}{A numeric. The sample size of external datasets}

\item{maf0}{A numeric. The MAF of the cases.}

\item{maf1}{A numeric. The MAF of the controls}

\item{maf.ext}{A numeric. The MAF of the external datasets.}

\item{pop.prev}{A numeric. The population prevalence of the disease.}

\item{var.ratio}{A numeric. The variance ratio calculated by sparseGRM.}
}
\value{
A numeric of batch effect p-value
}
\description{
This function test for the allele frequency difference between the study cohort and the external datasets.
}
