% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixPlus.R
\name{checkControl.Marker.SPAmixPlus}
\alias{checkControl.Marker.SPAmixPlus}
\title{Check control parameters for SPAmixPlus marker testing}
\usage{
checkControl.Marker.SPAmixPlus(control)
}
\description{
Check control parameters for SPAmixPlus marker testing
}
\keyword{internal}
