% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixPlus.R
\name{checkControl.NullModel.SPAmixPlus}
\alias{checkControl.NullModel.SPAmixPlus}
\title{Check control parameters for SPAmixPlus null model}
\usage{
checkControl.NullModel.SPAmixPlus(traitType, GenoFile, SparseGRMFile, control)
}
\description{
Check control parameters for SPAmixPlus null model
}
\keyword{internal}
