% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{checkOutputFile}
\alias{checkOutputFile}
\title{Check and validate output file settings for analysis restart capability}
\usage{
checkOutputFile(OutputFile, OutputFileIndex, AnalysisType, nEachChunk)
}
\description{
Check and validate output file settings for analysis restart capability
}
