% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixPlus.R
\name{fitNullModel.SPAmixPlus}
\alias{fitNullModel.SPAmixPlus}
\title{Fit null model for SPAmixPlus}
\usage{
fitNullModel.SPAmixPlus(response, designMat, subjData, control, SparseGRMFile)
}
\description{
Fit null model for SPAmixPlus
}
\keyword{internal}
