% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect.R
\name{getResid.wCox}
\alias{getResid.wCox}
\title{Get model residuals using a weighted Cox regression model.}
\usage{
getResid.wCox(formula, PhenoData, RefPrevalence)
}
\arguments{
\item{formula}{an R formula}

\item{PhenoData}{a dataframe including at least 3 columns c("SampleID","Indicator","time")}

\item{RefPrevalence}{consisitent with the RefPrevalence used in QCforBatchEffect.}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{mergeGenoInfo}: XXX.
\item{cutoff}: XXX.
\item{count}: XXX.
\item{PhenoData}: XXX.
\item{control}: XXX.
}
}
\description{
Get model residuals using a weighted Cox regression model.
}
