% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{GRAB.control}
\alias{GRAB.control}
\title{Information about the control in GRAB package}
\usage{
GRAB.control()
}
\description{
Information about the control in GRAB package
}
\details{
In GRAB package, we use \code{control} to specify parameters in multiple functions. Here, we only list generic parameters. 
For the parameters only used in specific method, please refer to its own help page. 
For example, check \code{?GRAB.SPACox} for more information about SPACox method.
\describe{
  \item{GRAB.ReadGeno}{
  We support two include files of (IDsToIncludeFile, RangesToIncludeFile) and two exclude files of (IDsToExcludeFile, RangesToExcludeFile), but do not support both include and exclude files.
  \itemize{
  \item \code{IDsToIncludeFile}: a file of marker IDs to include, one column (no header). Check \code{system.file("extdata", "IDsToInclude.txt", package = "GRAB")} for an example. 
  \item \code{IDsToExcludeFile}: a file of marker IDs to exclude, one column (no header). 
  \item \code{RangesToIncludeFile}: a file of ranges to include, three columns (no headers): chromosome, start position, end position. Check \code{system.file("extdata", "RangesToInclude.txt", package = "GRAB")} for an example.
  \item \code{RangesToExcludeFile}: a file of ranges to exclude, three columns (no headers): chromosome, start position, end position.
  \item \code{AlleleOrder}: a character, "ref-first" or "alt-first", to determine whether the REF/major allele should appear first or second. Default is "alt-first" for PLINK and "ref-first" for BGEN. If the analysis results show the ALT allele frequencies of most markers are > 0.5, you might should change this option. NOTE, if you use plink2 to convert Plink file to BGEN file, you probably need to set it as 'ref-first'.
  }
  }
  \item{GRAB.NullModel}{
  The following parameters are to handle dense GRM for mixed model approaches (\code{SAIGE}, \code{GATE}, and \code{POLMM})
    \itemize{
    \item \code{memoryChunk}: xxx
    \item \code{seed}: xxx
    \item \code{tracenrun}: xxx
    \item \code{maxiter}: xxx
    \item \code{tolBeta}: xxx
    \item \code{tolTau}: xxx
    \item \code{tau}: xxx
    \item \code{maxiterPCG}: xxx
    \item \code{tolEps}: xxx
    \item \code{minMafVarRatio}: xxx
    \item \code{maxMissingVarRatio}: xxx
    \item \code{nSNPsVarRatio}: xxx
    \item \code{CVcutoff}: xxx
    \item \code{LOCO}: xxx
    \item \code{numThreads}: xxx
    \item \code{stackSize}: xxx
    \item \code{grainSize}: xxx
    \item \code{minMafGRM}: xxx
    \item \code{maxMissingGRM}: xxx
    \item \code{showInfo }: xxx
    \item \code{onlyCheckTime}: xxx
    }
  }
  \item{GRAB.Marker}{
  \itemize{
  The following parameters are to specify Markers to analyze.
  \item \code{IDsToIncludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{IDsToExcludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{RangesToIncludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{RangesToExcludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{AlleleOrder}: please refer to section \code{GRAB.ReadGeno}.
  The following parameter are for analysis
  \item \code{omp_num_threads}: a numeric value (default: value from data.table::getDTthreads()) to specify the number of threads in OpenMP for parallel computation.
  \item \code{impute_method}: a character, "mean", "minor", or "drop". If "mean", impute genotype using 2 * AlleleFreq; if "minor", impute genotype using minor alleles; if "drop", drop the subject whose genotype is missing.
  \item \code{missing_cutoff}: a numeric value (default: 0.15). Any variant with missing rate > this value will be excluded from analysis.  
  \item \code{min_maf_marker}: a numeric value (default: 0.001). Any variants with MAF < this value will be excluded from analysis.  
  \item \code{min_mac_marker}: a numeric value (default: 20). Any variants with MAC < this value will be excluded from analysis.  
  \item \code{nMarkersEachChunk}: number of markers (default: 10000) in one chunk to be outputted
  }
  }
  \item{GRAB.Region}{
  \itemize{
  \item \code{IDsToIncludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{IDsToExcludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{RangesToIncludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{RangesToExcludeFile}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{AlleleOrder}: please refer to section \code{GRAB.ReadGeno}.
  \item \code{impute_method}: a character, "mean", "minor", or "drop". If "mean", impute genotype using 2 * AlleleFreq; if "minor", impute genotype using minor alleles; if "drop", drop the subject whose genotype is missing.
  \item \code{missing_cutoff}: a numeric value (default: 0.15). Any variant with missing rate > this value will be excluded from analysis.  
  \item \code{max_maf_region}: a numeric value (default: 0.001). Any variants with MAF < this value will be excluded from analysis.  
  \item \code{nRegionsEachChunk}: number of markers (default: 10000) in one chunk to be outputted
  }
  }
}
}
