% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtGRAB_Marker.R
\name{GRAB.mtMarker}
\alias{GRAB.mtMarker}
\title{GRAB.mtMarker}
\usage{
GRAB.mtMarker(
  objNull,
  GenoFile,
  OutputFile,
  GenoFileIndex = NULL,
  overwrite = FALSE,
  control = list()
)
}
\arguments{
\item{objNull}{A fitted null model object from \code{GRAB.NullModel}.}

\item{GenoFile}{Path to PLINK bed file.}

\item{OutputFile}{Path to output file for marker-level results. ".gz" extension for gzip compression.}

\item{GenoFileIndex}{Optional character vector of length 2 with paths to PLINK bim and fam files.}

\item{overwrite}{Whether to overwrite existing OutputFile (default FALSE).}

\item{control}{Set \code{nthreads} to specify the number of threads for parallel processing (default is the number of available cores minus one, up to a maximum of 8).}
}
\description{
Multithreaded GRAB.Marker
}
\examples{
extdir <- system.file("extdata", package = "GRABmtMarker")
GenoFile <- paste0(extdir, "/simuPLINK.bed")
OutputFile <- tempfile(fileext = ".tsv.gz")
control <- list(nMarkersEachChunk = 256, nthreads = 3)
supported_methods <- c("POLMM", "LEAF", "WtCoxG", "SPAGRM", "SAGELD", "SPAsqr", "SPACox", "SPAmix", "SPAmixPlus")

for (method in supported_methods) {
  objNull <- readRDS(paste0(extdir, "/obj.", method, ".rds"))
  if (method == "SPAmixPlus") control$afFilePath = paste0(extdir, "/afModels.bin")
  GRAB.mtMarker(objNull, GenoFile, OutputFile, control = control, overwrite = TRUE)
  head(read.table(OutputFile, header = TRUE))
}

}
