% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGRM.R
\name{GRAB.RPGWAS}
\alias{GRAB.RPGWAS}
\title{RPGWAS method in GRAB package}
\usage{
  GRAB.RPGWAS()
}
\description{
  RPGWAS method is an empirical approach to jointly analyzing binary trait and its risk prediction for related samples in a large-scale biobank.
}
\details{
  Additional list of \code{control} in \code{RPGWAS.NullModel()} function.
  
  Additional list of \code{control} in \code{GRAB.Marker()} function.
}
\examples{
  # Step 2a: process model residuals
  PhenoFile = system.file("RPGWAS", "simuPHENO.txt", package = "GRAB")
  PhenoData = data.table::fread(PhenoFile, header = T)
  
  PlinkFile = system.file("RPGWAS", "markers_for_RPGWAS_lambda.bed", package = "GRAB")
  ResidMatFile = system.file("extdata", "ResidMat.txt", package = "GRAB")
  SparseGRMFile = system.file("SparseGRM", "SparseGRM.txt", package = "GRAB")
  PairwiseIBDFile = system.file("PairwiseIBD", "PairwiseIBD.txt", package = "GRAB")
  obj.RPGWAS = RPGWAS.NullModel(PhenoFile,
                                TarColname = 'Tar',
                                RiskColname = 'Risk',
                                CovaColname = c('AGE', 'GENDER'),
                                PlinkFile = PlinkFile,
                                SparseGRMFile = SparseGRMFile,
                                PairwiseIBDFile = PairwiseIBDFile,
                                control = list(ControlOutlier = FALSE))
  
  # Step 2b: perform score test
  GenoFile = system.file("extdata", "simuPLINK.bed", package = "GRAB")
  OutputDir = system.file("RPGWAS", package = "GRAB")
  OutputFile = paste0(OutputDir, "/RPGWASMarkers.txt")
  GRAB.Marker(objNull = obj.RPGWAS,
              GenoFile = GenoFile,
              OutputFile = OutputFile)
  head(read.table(OutputFile, header=T))
}