% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Simu.R
\name{GRAB.SimuGMatSKAT}
\alias{GRAB.SimuGMatSKAT}
\title{GRAB: simulate genotype matrix based on family structure}
\usage{
GRAB.SimuGMatSKAT(
  n.fam,
  n.subj,
  fam.kin.mode,
  min.MAF,
  max.MAF,
  seedNum,
  region.length = 2000,
  n.region,
  prob = 1
)
}
\arguments{
\item{n.fam}{number of families in simulation}

\item{n.subj}{number of unrelated subjects in simulation}

\item{fam.kin.mode}{"4-members" or "10-members" to simulate two family structures. "10-members" corresponds to system.file("extdata", "example_10members.kin.txt", package = "GRAB")}

\item{min.MAF}{minimal MAF in simulation}

\item{max.MAF}{maximal MAF in simulation}

\item{seedNum}{seed number for random simulation}

\item{region.length}{length of region to simulate. Default value is 2000, i.e., 2KB.}

\item{n.region}{number of regions to simulate}
}
\value{
a matrix of genotype data including multiple Genes
}
\description{
Simulate genotype matrix based on family structure using Haplotype stored in SKAT package
}
\examples{
fam.kin.10.members = read.table(system.file("extdata", "example_10members.kin.txt", package = "GRAB"))
fam.kin.10.members = as.matrix(fam.kin.10.members)
n.fam = 500
n.subj = 5000
fam.kin.mode = "10-members"  # "10-members" corresponds to fam.kin.10.members
min.MAF = 0
max.MAF = 0.3
seedNum = 1234
region.length = 2000
n.region = 2
GMat = GRAB.SimuGMatSKAT(n.fam, n.subj, fam.kin.mode, min.MAF, max.MAF, seedNum, region.length, n.region)

}
