% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmixLocalPlus.R
\name{GRAB.SPAmixLocalPlus}
\alias{GRAB.SPAmixLocalPlus}
\title{Instruction of SPAmixLocalPlus method}
\usage{
GRAB.SPAmixLocalPlus()
}
\description{
SPAmixLocalPlus performs local ancestry-specific association testing by accounting
for local ancestry patterns using ancestry-specific haplotype counts and pre-computed
phi (intuitively kinship coefficients stratified by local ancestry) matrices.
}
\examples{
SparseGRMFile <- system.file("extdata", "SparseGRM.txt", package = "GRAB")
GRM <- data.table::fread(SparseGRMFile)

PhenoFile <- system.file("extdata", "simuPHENO.txt", package = "GRAB")
PhenoData <- data.table::fread(PhenoFile)
subjData <- PhenoData$IID

extdata_dir <- system.file("extdata", package = "GRAB")
dosagePrefix <- paste0(extdata_dir, "/simuAncestry")

temp_dir <- tempdir()
phiOutputPrefix <- paste0(temp_dir, "/phi_")
outPrefix <- paste0(temp_dir, "/out_")

# Step0: Estimate phi for ancestries 1 and 2
SPAmixLocalPlus.EstimatePhi(
GRM = GRM,
  dosagePrefix = dosagePrefix,
  haploPrefix = NULL,  # Uses dosagePrefix if NULL
  ancIdx = c(1, 2),
  SampleIDs = subjData,
  phiOutputPrefix = phiOutputPrefix,
  MAF_cutoff = 0.01
)

# Step1: Fit null model to get residuals
residuals <- survival::coxph(
  survival::Surv(SurvTime, SurvEvent) ~ AGE + GENDER + PC1 + PC2,
  data = PhenoData
)$residuals

# Step2: Run SPAmixPlus tests for ancestries 1 and 2
result_lst <- SPAmixLocalPlus.Marker(
  resid = residuals,
  subjData = subjData,
  dosagePrefix = dosagePrefix,
  haploPrefix = NULL,  # Uses dosagePrefix if NULL
  phiPrefix = phiOutputPrefix,
  ancIdx = c(1, 2),
  outPrefix = outPrefix
)

# View results
OutputFile1 <- paste0(outPrefix, "Ancestry1.txt")
head(data.table::fread(OutputFile1))

}
\references{
Yuzhuo Ma (2026). \url{https://github.com/YuzhuoMa97/SPAmixPlus}
}
\seealso{
\code{\link{SPAmixLocalPlus.EstimatePhi}} for phi matrix estimation,
\code{\link{SPAmixLocalPlus.Marker}} for the main analysis function
}
