% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtSPAG.R
\name{GRAB.WtSPAG}
\alias{GRAB.WtSPAG}
\title{WtSPAG method in GRAB package}
\usage{
GRAB.WtSPAG()
}
\description{
WtSPAG method is an empirical approach to analyzing complex traits (including but not limited to time-to-event trait) for unrelated samples in a large-scale biobank.
}
\details{
Additional list of \code{control} in \code{GRAB.NullModel()} function.

Additional list of \code{control} in \code{GRAB.Marker()} function.
}
\examples{
setwd(system.file("WtSPAG", package = "GRAB"))
PhenoData = read.table(system.file("WtSPAG", "simuPHENO_WtSPAG.txt", package = "GRAB"), header = T)
RefPrevalence = 0.1

qcResult1 = QCforBatchEffect(GenoFile = "simuBGEN1.bgen", 
                             GenoFileIndex = c("simuBGEN1.bgen.bgi", 
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first", 
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoData=PhenoData,
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             SNPnum=1e4)

qcResult2 = QCforBatchEffect(GenoFile = "GenoMat2.bed",
                             OutputFile = "qcBGEN2.txt",
                             control=list(AlleleOrder = "ref-first", 
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoData=PhenoData,
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             SNPnum=1e4)

allQC = mergeQCresults(qcResult1, qcResult2)
 
obj.WtSPAG = GRAB.NullModel(Surv(SurvTime, SurvEvent) ~ Cov1 + Cov2,
                           data = PhenoData, 
                           subjData = IID, 
                           method = "WtSPAG", 
                           traitType = "time-to-event",
                           control = list(RefPrevalence = RefPrevalence))

names(obj.WtSPAG)

GRAB.Marker(obj.WtSPAG, 
            GenoFile = "simuBGEN1.bgen",
            GenoFileIndex = c("simuBGEN1.bgen.bgi", "simuBGEN1.sample"),
            OutputFile = "simuBGEN1.txt", 
            OutputFileIndex = NULL, 
            control = list(QCFile = "qcBGEN1.txt",
                           QCCutoff = qcResult1$cutoff,
                           AlleleOrder = "ref-first", 
                           AllMarkers = T))

}
