% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Null_Model.R
\name{GRAB_Null_Model}
\alias{GRAB_Null_Model}
\title{Fits a NULL model for SPACox}
\usage{
GRAB_Null_Model(
  formula,
  data = NULL,
  subjData,
  method = "SPACox",
  trait.type = "time-to-event",
  GenoFile,
  GeniFileIndex = NULL,
  control = NULL,
  range = c(-100, 100),
  length.out = 10000,
  ...
)
}
\arguments{
\item{formula}{a formula to be passed to function coxph(). For more details, please refer to package survival.}

\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{subjData}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order in the formula.}

\item{method}{an R character, "SPACox", "SPAGE", "SAIGE", "POLMM", "GATE"}

\item{trait.type}{an R character, "binary", "categorical", "quantitative", "time-to-event"}

\item{GenoFile}{xxxxxxxxxxx}

\item{control}{xxxxxxxxxxx}

\item{range}{a two-element numeric vector (default: c(-100,100)) to specify the domain of the empirical CGF.}

\item{length.out}{a positive integer (default: 9999) for empirical CGF. Larger length.out corresponds to longer calculation time and more accurate estimated empirical CGF.}

\item{...}{Other arguments passed to function coxph(). For more details, please refer to package survival.}

\item{GenoFileIndex}{xxxxxxxxxxx}
}
\value{
an object with a class of "SPACox_NULL_Model".
}
\description{
Fits a null Cox proportional hazards model and then calculates the empirical cumulant generation function (CGF) of the martingale residuals
}
\examples{
Please check help(SPACox) for a simulated example.
}
