% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect.R
\name{mergeQCresults}
\alias{mergeQCresults}
\title{Merge quality control results from multiple genotype files.}
\usage{
mergeQCresults(qcResult1, ...)
}
\arguments{
\item{qcResult1}{the output R list of function QCforBatchEffect().}

\item{...}{other output R lists of function QCforBatchEffect().}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{mergeGenoInfo}: XXX.
\item{cutoff}: XXX.
\item{count}: XXX.
\item{PhenoData}: XXX.
\item{control}: XXX.
}
}
\description{
Merge quality control results from multiple genotype files.
}
