% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Simu.R
\name{plink.make}
\alias{plink.make}
\title{GRAB: make plink files using Geno.mat}
\usage{
plink.make(Geno.mat, work.dir, out.prefix, mCHRs = T, BP = NULL)
}
\arguments{
\item{Geno.mat}{Genotype matrix with row names of subject IDs and column names of SNP IDs}

\item{work.dir}{working directory to store the plink files}

\item{out.prefix}{prefix of the output Plink files}

\item{mCHRs}{if TRUE, then SNPs were randomly assigned to chromosomes 1-22; if FALSE, then all SNPs are in chromosome 1}

\item{BP}{if NULL, then Base Position (BP) is 1:length(SNPs), otherwise, the length(BP) should be the same as ncol(Geno.mat)}
}
\value{
Plink files are stored in 'work.dir' with prefix of 'out.prefix'.
}
\description{
Make plink files using Geno.mat, rownames(Geno.mat) should be subject IDs and colnames(Geno.mat) should be SNP IDs
}
\examples{
n = 1000
m = 20
MAF = 0.3
Geno.mat = matrix(rbinom(n*m, 2, MAF), n, m)
rownames(Geno.mat) = paste0("subj-",1:n)
colnames(Geno.mat) = paste0("SNP-",1:m)
work.dir = system.file("results", package = "GRAB")
out.prefix = "test"
mCHRs = FALSE
BP = NULL
plink.make(Geno.mat, work.dir, out.prefix, mCHRs, BP)
     
# cmd = paste0("plink --file ", OUT.file,  " --make-bed --out ", OUT.file)
# system(cmd)
# system(paste("rm", MAP.file, PED.file))
     
}
