% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLMM.R
\name{setMarker.POLMM}
\alias{setMarker.POLMM}
\title{Set up marker-level testing for POLMM method}
\usage{
setMarker.POLMM(objNull, control, chrom)
}
\arguments{
\item{objNull}{Null model object from GRAB.NullModel().}

\item{control}{List of control parameters.}

\item{chrom}{Character string specifying chromosome for LOCO analysis.}
}
\value{
List containing setup parameters for marker testing.
}
\description{
Set up marker-level testing for POLMM method
}
\keyword{internal}
