% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGELD.R
\name{setMarker.SAGELD}
\alias{setMarker.SAGELD}
\title{Set up marker-level testing for SAGELD method}
\usage{
setMarker.SAGELD(objNull, control)
}
\arguments{
\item{objNull}{Null model object from SAGELD.NullModel().}

\item{control}{List of control parameters.}
}
\value{
List containing setup parameters for marker testing.
}
\description{
Set up marker-level testing for SAGELD method
}
\keyword{internal}
