% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPACox.R
\name{setMarker.SPACox}
\alias{setMarker.SPACox}
\title{Set up marker-level testing for SPACox method}
\usage{
setMarker.SPACox(objNull, control)
}
\arguments{
\item{objNull}{Null model object from GRAB.NullModel().}

\item{control}{List of control parameters.}
}
\value{
List containing setup parameters for marker testing.
}
\description{
Set up marker-level testing for SPACox method
}
\keyword{internal}
