% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmix.R
\name{setMarker.SPAmix}
\alias{setMarker.SPAmix}
\title{Set up marker-level testing for SPAmix method}
\usage{
setMarker.SPAmix(objNull, control)
}
\arguments{
\item{objNull}{Null model object from GRAB.NullModel().}

\item{control}{List of control parameters.}
}
\value{
List containing setup parameters for marker testing.
}
\description{
Set up marker-level testing for SPAmix method
}
\keyword{internal}
