% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLMM.R
\name{setRegion.POLMM}
\alias{setRegion.POLMM}
\title{Set up region-based testing for POLMM method}
\usage{
setRegion.POLMM(objNull, control, chrom, SparseGRMFile)
}
\arguments{
\item{objNull}{Null model object from GRAB.NullModel().}

\item{control}{List of control parameters.}

\item{chrom}{Character string specifying chromosome for LOCO analysis.}

\item{SparseGRMFile}{Character string specifying sparse GRM file path.}
}
\value{
List containing setup parameters for region testing.
}
\description{
Set up region-based testing for POLMM method
}
\keyword{internal}
