% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{writeOutputFile}
\alias{writeOutputFile}
\title{Write analysis output to files with progress tracking}
\usage{
writeOutputFile(
  Output,
  OutputFile,
  OutputFileIndex,
  AnalysisType,
  nEachChunk,
  indexChunk,
  Start,
  End
)
}
\arguments{
\item{Output}{List of output data to write to files.}

\item{OutputFile}{Character vector of output file paths.}

\item{OutputFileIndex}{Character string specifying the index file path.}

\item{AnalysisType}{Character string indicating analysis type.}

\item{nEachChunk}{Integer specifying the number of items per chunk.}

\item{indexChunk}{Integer indicating the current chunk index.}

\item{Start}{Logical indicating if this is the first output to save.}

\item{End}{Logical indicating if this is the last output to save.}
}
\description{
Write analysis output to files with progress tracking
}
