% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPACox.R
\name{fitNullModel.SPACox}
\alias{fitNullModel.SPACox}
\title{Fit null model for SPACox method}
\usage{
fitNullModel.SPACox(response, designMat, subjData, control, ...)
}
\arguments{
\item{response}{Response variable (Surv or Residual object).}

\item{designMat}{Design matrix containing covariates.}

\item{subjData}{Vector of subject identifiers.}

\item{control}{List of control parameters.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
List containing fitted null model components.
}
\description{
Fit null model for SPACox method
}
\keyword{internal}
