% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtSPAG.R
\name{getResid.wCox}
\alias{getResid.wCox}
\title{Get model residuals using a weighted Cox regression model.}
\usage{
getResid.wCox(formula, PhenoData, RefPrevalence)
}
\arguments{
\item{formula}{an refression formula: a symbolic description of the null model to be fitted}

\item{PhenoData}{a dataframe including at least 3 columns \code{SampleID}, \code{Indicator} and \code{time}}

\item{RefPrevalence}{a numeric of the event rate in the population, which is consisitent with the input in QCforBatchEffect.}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{Residual}: A dataframe with 2 columns, \code{SampleID} and \code{Residuals}, and \code{Residuals} are calculated by fitting null model.
}
}
\description{
Get model residuals using a weighted Cox regression model.
}
