% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno.R
\name{getVersionFromBGEN}
\alias{getVersionFromBGEN}
\title{Get version information from BGEN file}
\usage{
getVersionFromBGEN(bgenFile)
}
\arguments{
\item{bgenFile}{a character of BGEN file.}
}
\description{
Get version information from BGEN file (check \href{https://www.well.ox.ac.uk/~gav/bgen_format/spec/v1.2.html}{link})
}
\examples{

BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "GRAB")
getVersionFromBGEN(BGENFile)
}
