% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Marker.R
\name{GRAB.Marker}
\alias{GRAB.Marker}
\title{GRAB: Marker-level score test}
\usage{
GRAB.Marker(
  objNull,
  GenoFile,
  GenoFileIndex = NULL,
  OutputFile,
  OutputFileIndex = NULL,
  control = NULL
)
}
\arguments{
\item{objNull}{output object of the GRAB.NullModel() function.}

\item{GenoFile}{a character of genotype file. Three types of genotype files are supported: PLINK ("prefix.bed"), BGEN ("prefix.bgen"), and VCF ("prefix.vcf" or "prefix.vcf.gz").}

\item{GenoFileIndex}{additional index files corresponding to the "GenoFile". If Null (default), the same prefix as GenoFile is used. PLINK: c("prefix.bim", "prefix.fam"), BGEN: c("prefix.bgi"), and VCF: c("prefix.vcf.tbi") or c("prefix.vcf.gz.tbi").}

\item{OutputFile}{a character of output file to store the analysis results.}

\item{OutputFileIndex}{a character of output index file to record the end point in case that program ends unexpectedly}

\item{control}{a list of parameters for controlling the GRAB.Marker(), more details can be found in \code{?GRAB.control}.}
}
\value{
The results will be written in a file, i.e., OutputFile, which includes the following columns.
\item{Marker}{Marker IDs extracted from "GenoFile" and "GenoFileIndex".}
\item{Info}{Marker Infomation of "CHR:POS:REF:ALT". This information is from "GenoFile" or "GenoFileIndex". Note that control of \code{AlleleOrder}, i.e. "ref-first" or "alt-first" may alter the order.}
\item{AltFreq}{ALT allele frequency (mighe be > 0.5). If most of the AltFreq are > 0.5, you might should reset the control of \code{AlleleOrder}. Refer to section of \code{GRAB.ReadGeno} in \code{?GRAB.control} for more details.}
\item{AltCounts}{ALT allele counts.}
\item{MissingRate}{Missing rate of marker}
\item{Beta}{Estimated effect size (if provided), of the ALT allele.}
\item{seBeta}{Estimated standard error (se, if provided) of the effect size}
\item{Pval}{Association test p-value}
\item{zScore}{z value (if provided), standardized score statistics, usually follows a standard normal distribution}
}
\description{
GRAB package uses score test for GWAS: in step 1, we fit a null model (check \code{?GRAB.NullModel}) including response variable, covariates, and GRM (if needed). In step 2, we perform score test for marker-level analysis (check \code{?GRAB.Marker}) and region-level analysis (check \code{?GRAB.Region}).
}
\details{
Check \code{?GRAB.control} for a generic list of 'control'. For specific methods, check its help page for more information. 
\itemize{
\item{SAIGE: to analyze binary phenotype, check ?GRAB.SAIGE for more details.}
\item{POLMM: to analyze ordinal categorical phenotype, check ?GRAB.POLMM for more details.}
\item{SPACox: to analyze time-to-event phenotype, check ?GRAB.SPACox for more details.}
\item{SPAGE: to analyze gene-environment interaction effect for binary phenotype, check ?GRAB.SPAGE for more details.}
}
}
\examples{
# We put examples to the specific help pages for different methods. 
# If you want to use "SPACox" method, please check ?GRAB.SPACox for more details.
# If you want to use "POLMM" method, please check ?GRAB.POLMM for more details.
}
