% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Marker.R
\name{GRAB.Marker}
\alias{GRAB.Marker}
\title{GRAB: Marker-level genetic analysis}
\usage{
GRAB.Marker(
  objNull,
  GenoFile,
  GenoFileIndex = NULL,
  OutputFile = NULL,
  control = NULL
)
}
\arguments{
\item{objNull}{output object of the GRAB.NullModel() function.}

\item{GenoFile}{a character of genotype file. Three types of genotype files are supported: PLINK ("prefix.bed"), BGEN ("prefix.bgen"), and VCF ("prefix.vcf" or "prefix.vcf.gz").}

\item{GenoFileIndex}{additional index files corresponding to the "GenoFile". If Null (default), the same prefix as GenoFile is used. PLINK: c("prefix.bim", "prefix.fam"), BGEN: c("prefix.bgi"), and VCF: c("prefix.vcf.tbi") or c("prefix.vcf.gz.tbi").}

\item{OutputFile}{a character of output file to store the analysis results. If Null (default), the results are outputted to R.}

\item{control}{a list of parameters for controlling the GRAB.Marker().}
}
\value{
The results will be written in a file (if OutputFile != NULL) or be saved to an R data.frame (default, if OutputFile == NULL). The results include the following columns.
\item{Marker}{Marker IDs extracted from "GenoFile" and "GenoFileIndex".}
\item{Info}{Marker Infomation of "CHR:POS:REF:ALT". This information is from "GenoFile" or "GenoFileIndex" and does not change even if the REF/ALT alleles are flipped in analysis.}
\item{Freq}{Minor allele frequency (always < 0.5) in analysis.}
\item{Flip}{Logical value indicating if the REF/ALT alleles were switched in analysis. This information is useful to estimate the effect direction.}
\item{Beta}{Estimated effect size. The sign (positive or negative) depends on "Info" and "Flip". For example, if 'Flip' is false and 'Beta' is positive, then ALT allele is to increase continuous trait (trait.type == "quantitative"), to increase the risk of being cases (trait.type == "binary"), or to xxxx}
\item{seBeta}{Estimated standard error (se) of the effect size}
\item{Pval}{p-value from normal distribution approximation or saddlepoint approximation.}
}
\description{
Marker-level genetic analysis: test for association between a phenotype of interest and genome-wide genetic markers
}
\details{
List of 'control' is different for different methods
\itemize{
\item{SAIGE: to analyze binary phenotype, check ?GRAB.SAIGE for more details.}
\item{POLMM: to analyze ordinal categorical phenotype, check ?GRAB.POLMM for more details.}
\item{SPACox: to analyze time-to-event phenotype, check ?GRAB.SPACox for more details.}
\item{SPAGE: to analyze gene-environment interaction effect for binary phenotype, check ?GRAB.SPAGE for more details.}
}
}
\examples{
# We put examples to the specific help pages for different methods. 
# For example, if you want to use "SPACox" method, please check ?GRAB.SPACox for more details. 
}
