% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Null_Model.R
\name{GRAB.NullModel}
\alias{GRAB.NullModel}
\title{Fits a NULL model for SPACox}
\usage{
GRAB.NullModel(
  formula,
  data = NULL,
  subjData,
  method = "SPACox",
  trait.type = "time-to-event",
  GenoFile,
  GenoFileIndex = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula to be passed to function coxph(). For more details, please refer to package survival.}

\item{data}{a data.frame in which to interpret the variables named in the formula}

\item{subjData}{a character vector of subject IDs. NOTE: its order should be the same as the subjects order in the formula.}

\item{method}{an R character, "SPACox", "SPAGE", "SAIGE", "POLMM", "GATE"}

\item{trait.type}{an R character, "binary", "categorical", "quantitative", "time-to-event"}

\item{GenoFile}{xxxxxxxxxxx}

\item{GenoFileIndex}{xxxxxxxxxxx}

\item{control}{xxxxxxxxxxx}

\item{...}{Other arguments passed to function coxph(). For more details, please refer to package survival.}

\item{range}{a two-element numeric vector (default: c(-100,100)) to specify the domain of the empirical CGF.}

\item{length.out}{a positive integer (default: 9999) for empirical CGF. Larger length.out corresponds to longer calculation time and more accurate estimated empirical CGF.}
}
\value{
an object with a class of "SPACox_NULL_Model".
}
\description{
Fits a null Cox proportional hazards model and then calculates the empirical cumulant generation function (CGF) of the martingale residuals
}
\examples{
# Simulation phenotype and genotype
GenoFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.bed", package = "GRAB")
N = 100
Pheno = data.frame(ID = paste0("f",1:N,"_1"),
                   event=rbinom(N,1,0.5),
                   time=runif(N),
                   Cov1=rnorm(N),
                   Cov2=rbinom(N,1,0.5))
obj.SPACox = GRAB.NullModel(survival::Surv(time,event)~Cov1+Cov2, 
                            data=Pheno, subjData = Pheno$ID, method = "SPACox", GenoFile = GenoFile)
}
