% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Null_Model.R
\name{GRAB.NullModel}
\alias{GRAB.NullModel}
\title{Fits a NULL model}
\usage{
GRAB.NullModel(
  formula,
  data = NULL,
  subset = NULL,
  subjData,
  method = "SPACox",
  trait.type = "time-to-event",
  GenoFile,
  GenoFileIndex = NULL,
  control = NULL,
  ...
)
}
\arguments{
\item{formula}{a formula object, with the response on the left of a ~ operator and the covariates on the right. Do not add a column of intercept (e.g. a vector of ones) on the right. Missing values should be denoted by NA and the corresponding samples will be removed from analysis.}

\item{data}{a data.frame in which to interpret the variables named in the formula, or in the subset argument. Check ?model.frame for more details.}

\item{subset}{a specification of the rows to be used: defaults to all rows. This can be any valid indexing vector for the rows of data or if that is not supplied, a data frame made up of the variables used in formula. Check ?model.frame for more details.}

\item{subjData}{a character vector of subject IDs. Its order should be the same as the subjects order in the formula and data.}

\item{method}{an R character, "SPACox", "SPAGE", "SAIGE", "POLMM", or "GATE"}

\item{trait.type}{an R character, "binary", "categorical", "quantitative", or "time-to-event"}

\item{GenoFile}{a character of genotype file. Three types of genotype files are supported: PLINK ("prefix.bed"), BGEN ("prefix.bgen"), and VCF ("prefix.vcf" or "prefix.vcf.gz").}

\item{GenoFileIndex}{additional index files corresponding to the "GenoFile". If Null (default), the same prefix as GenoFile is used. PLINK: c("prefix.bim", "prefix.fam"), BGEN: c("prefix.bgi"), and VCF: c("prefix.vcf.tbi") or c("prefix.vcf.gz.tbi").}

\item{control}{a list of parameters for controlling the GRAB.NullModel().}

\item{...}{Other arguments passed to function XXXX().}
}
\value{
an R object with a class of "XXXXX_NULL_Model" in which XXXXX is the 'method' used in analysis.
}
\description{
Fits a null model
}
\examples{
We put examples to the specific help pages for different methods. 
For example, if you want to use "SPACox" method, please check ?GRAB.SPACox for more details. 
}
