% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POLMM.R
\name{GRAB.POLMM}
\alias{GRAB.POLMM}
\title{How to run POLMM method in GRAB package}
\usage{
GRAB.POLMM()
}
\description{
POLMM method is to analysis ordinal categorical phenotype for large-scale biobank data
}
\details{
List of 'control' in GRAB.NullModel function
\itemize{
\item{range: a two-element numeric vector [default=c(-100,100)] to specify the domain of the empirical CGF.}
\item{length.out: a positive integer [default=10000] for empirical CGF. Larger length.out corresponds to longer calculation time and more accurate estimated empirical CGF.}
}
List of 'control' in GRAB.Marker function
\itemize{
\item{impute_method: a character string [default="fixed"] to specify the method to impute missing genotypes. "fixed" imputes missing genotypes (NA) by assigning the mean genotype value (i.e. 2p where p is MAF).}
\item{missing_cutoff: a numeric value [default=0.15] to specify the cutoff of the missing rates. Any variant with missing rate higher than this cutoff will be excluded from the analysis.}
\item{min_maf_marker: a numeric value [default=0.001] to specify the cutoff of the minimal MAF. Any SNP with MAF < cutoff will be excluded from the analysis.}
\item{min_mac_marker: a numeric value [default=20] to specify the cutoff of the minimal MAC. Any SNP with MAC < cutoff will be excluded from the analysis.}
\item{nMarkersEachChunk: a numeric value [default=10000] to specify the number of markers in each chunk.}
\item{pVal_covaAdj_Cutoff: a numeric value [default=5e-5]. If the p-value is less than this cutoff, then we would use an additional technic to adjust for covariates.}
\item{SPA_cutoff: a numeric value [default=2] to specify the standard deviation cutoff to be used. If the test statistic lies within the standard deviation cutoff, its p value is calculated based on a normal distribution approximation, otherwise, its p value is calculated based on a saddlepoint approximation.".}
}
}
\examples{
# Simulation phenotype and genotype
GenoFile = system.file("extdata", "example.bed", package = "GRAB")
N = 100
Pheno = data.frame(ID = paste0("f",1:N,"_1"),
                   event=rbinom(N,1,0.5),
                   time=runif(N),
                   Cov1=rnorm(N),
                   Cov2=rbinom(N,1,0.5))
obj.SPACox = GRAB.NullModel(survival::Surv(time,event)~Cov1+Cov2, 
                            data=Pheno, subjData = Pheno$ID, method = "SPACox", GenoFile = GenoFile)

res.SPACox = GRAB.Marker(obj.SPACox, GenoFile)
head(res.SPACox)
}
