% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geno.R
\name{GRAB.ReadGeno}
\alias{GRAB.ReadGeno}
\title{Read in genotype data}
\usage{
GRAB.ReadGeno(
  GenoFile,
  GenoFileIndex = NULL,
  SampleIDs = NULL,
  control = NULL,
  sparse = FALSE
)
}
\arguments{
\item{GenoFile}{a character of genotype file. See \code{Details} section for more information.}

\item{GenoFileIndex}{additional index file(s) corresponding to the \code{GenoFile}. See \code{Details} section for more information.}

\item{SampleIDs}{a character vector of sample IDs to extract. The default is NULL, that is, to use all samples in GenoFile.}

\item{control}{a list of parameters to decide which markers to extract. If not specified, the first 10 markers will be extracted. For more details, please check \code{?GRAB.control}.}

\item{sparse}{a logical value (default: FALSE) to indicate if sparse genotype matrix is outputted.}
}
\value{
An R list include an R genotype matrix (each row is for one sample and each column is for one marker) and an R SNP information matrix.
}
\description{
GRAB package provides functions to read in genotype data from multiple format (including Plink, BGEN, and VCF) into R
}
\details{
We support three genotype formats including Plink, BGEN, and VCF. 
Users do not need to specify the genotype format, GRAB package will check the filename extention for that purpose.  
If \code{GenoFileIndex} is not specified, then GRAB uses the same prefix as \code{GenoFile}.
\describe{
  \item{Plink}{
  \itemize{
  \item \code{GenoFile}: "prefix.bed". 
  \item \code{GenoFileIndex}: c("prefix.bim", "prefix.fam").
  }
  }
  \item{BGEN}{
    \itemize{
    \item \code{GenoFile}: "prefix.bgen"; 
    \item \code{GenoFileIndex}: "prefix.bgen.bgi" or c("prefix.bgen.bgi", "prefix.bgen.samples").
    \item IMPORTANT NOTE: If only one element is given for \code{GenoFileIndex}, then we assume it should be "prefix.bgen.bgi". 
    If BGEN file does not include sample identifiers, then "prefix.bgen.samples" is required, which should be a file with only one column whose column name is "GRAB_BGEN_SAMPLE" (case insensitive). 
    One example is \code{system.file("extdata", "example_bgen_1.2_8bits.bgen.samples", package = "GRAB")}.
    If you are not sure if sample identifiers are in BGEN file, you can try function \code{?checkIfSampleIDsExist}.
    }
  }
  \item{VCF}{Not available now. \code{GenoFile}: "prefix.vcf"; \code{GenoFileIndex}: "prefix.vcf.tbi"}
}
}
\examples{

## The below is raw data 
RawFile = system.file("extdata", "example.raw", package = "GRAB")
GenoMat = data.table::fread(RawFile)
head(GenoMat[,1:15])

## The below is for Plink format
PlinkFile = system.file("extdata", "example.bed", package = "GRAB")
GenoList = GRAB.ReadGeno(PlinkFile)
GenoMat = GenoList$GenoMat
markerInfo = GenoList$markerInfo
head(GenoMat)
markerInfo

## The below is for BGEN format (Note the different REF/ALT order for BGEN and Plink formats)
BGENFile = system.file("extdata", "example_bgen_1.2_8bits.bgen", package = "GRAB")
GenoList = GRAB.ReadGeno(BGENFile)
GenoMat = GenoList$GenoMat
markerInfo = GenoList$markerInfo
head(GenoMat)
markerInfo

## The below is to demonstrate parameters in control
PlinkFile = system.file("extdata", "example.bed", package = "GRAB")
IDsToIncludeFile = system.file("extdata", "example.IDsToIncludeFile.txt", package = "GRAB")
RangesToIncludeFile = system.file("extdata", "example.RangesToIncludeFile.txt", package = "GRAB")
GenoList = GRAB.ReadGeno(PlinkFile, 
                         control = list(IDsToIncludeFile = IDsToIncludeFile, 
                                        RangesToIncludeFile = RangesToIncludeFile,
                                        AlleleOrder = "ref-first"))
GenoMat = GenoList$GenoMat
head(GenoMat)
markerInfo = GenoList$markerInfo
markerInfo

}
