% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Region.R
\name{GRAB.Region}
\alias{GRAB.Region}
\title{GRAB: region-based multiple markers analysis}
\usage{
GRAB.Region(
  objNull,
  GenoFile,
  GenoFileIndex = NULL,
  OutputFile,
  OutputFileIndex = NULL,
  RegionFile,
  RegionAnnoHeader = NULL,
  SparseGRMFile = NULL,
  control = NULL
)
}
\arguments{
\item{objNull}{an output object of the POLMM_Null_Model() function with a class of "POLMM_NULL_Model".}

\item{GenoFile}{a character of genotype file. Three types of genotype files are supported: PLINK ("prefix.bed"), BGEN ("prefix.bgen"), and VCF ("prefix.vcf" or "prefix.vcf.gz").}

\item{GenoFileIndex}{additional index files corresponding to the "GenoFile". The default is NULL, that is, to share the same prefix as GenoFile. PLINK: c("prefix.bim", "prefix.fam"), BGEN: c("prefix.bgi"), and VCF: c("prefix.vcf.tbi") or c("prefix.vcf.gz.tbi").}

\item{OutputFile}{a character of output file to store the analysis results}

\item{RegionFile}{a character of annotation file. Column 1: region ID, Column 2: marker ID, Columns 3-n annotation (non-negative) similar as in STAAR. The header is required and the first two should be "REGION" and "MARKER".}

\item{RegionAnnoHeader}{a character vector of annotation in analysis. Optional, if not specified, all annotation columns will be used.}

\item{control}{a character to specify chromosome of the markers in analysis. Must be specified unless LOCO = F when fitting the null model.}
}
\value{
an R matrix with the following elements
\item{ID}{Marker IDs from colnames(GMat)}
\item{chr}{Chromosome name from chrVec}
\item{MAF}{MAFs of the markers}
\item{missing.rate}{Missing rates of the markers}
\item{Stat}{Score statistics}
\item{VarW}{Estimated variance (VarW) from non-mixed model}
\item{VarP}{Estimated variance after adjusting for variance ratio r (VarP = VarW * r)}
\item{beta}{Estimated effect size: Stat / VarP}
\item{pval.norm}{p values calculated from normal approximation}
\item{pval.spa}{p values calculated from saddlepoint approximation}
\item{switch.allele}{a logical value indicating if the REF/ALT alleles were switched, if AF > 0.5, we use GVec = 2-GVec, and then give switch.allele=T. This is useful to estimate the effect direction.}
}
\description{
Region-based analysis: Test for association between multiple variants (mostly low-frequency or rare variants) in a region and an ordinal categorical phenotype via POLMM
}
\details{
More information about the list of 'SKAT.control'
\itemize{
\item{memory_chunk: a cutoff (Gb) to determine how many markers are in one chunk for region-based analysis [default=4].}
\item{kernel: how to weight markers for region-based analysis [default="linear.weighted"].}
\item{method: method to conduct region-based analysis [default="method"].}
\item{weights_beta:  [default=c(1,25)].}
\item{weights:  [default=NULL].}
\item{r_corr:  [default=NULL].}
\item{printPCGInfo:  [default=FALSE].}
\item{tolPCG:  [default=1e-5].}
\item{maxiterPCG:  [default=100].}
}
}
\examples{
# We put examples to the specific help pages for different methods. 
# If you want to use "POLMM" method, please check ?GRAB.POLMM for more details.
}
