% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGELD.R
\name{GRAB.SAGELD}
\alias{GRAB.SAGELD}
\title{SAGELD method in GRAB package}
\usage{
GRAB.SAGELD()
}
\description{
SAGELD method is Scalable and Accurate algorithm for Gene-Environment interaction analysis using Longitudinal Data for related samples in a large-scale biobank. SAGELD extended SPA\if{html}{\out{<sub>}}GRM\if{html}{\out{</sub>}} to support gene-environment interaction analysis.
}
\details{
Additional list of \code{control} in \code{SAGELD.NullModel()} function.

Additional list of \code{control} in \code{GRAB.Marker()} function.
}
