% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAGELD.R
\name{GRAB.SAGELD}
\alias{GRAB.SAGELD}
\title{SAGELD method in GRAB package}
\usage{
GRAB.SAGELD()
}
\description{
SAGELD method is Scalable and Accurate algorithm for Gene-Environment interaction analysis using Longitudinal Data for related samples in a large-scale biobank. SAGELD extended SPA\if{html}{\out{<sub>}}GRM\if{html}{\out{</sub>}} to support gene-environment interaction analysis.
}
\details{
Additional list of \code{control} in \code{SPAGRM.NullModel()} function.

Additional list of \code{control} in \code{GRAB.Marker()} function.
}
\examples{
# Step 2a: process model residuals
ResidMatFile = system.file("extdata", "ResidMat.txt", package = "GRAB")
SparseGRMFile = system.file("SparseGRM", "SparseGRM.txt", package = "GRAB")
PairwiseIBDFile = system.file("PairwiseIBD", "PairwiseIBD.txt", package = "GRAB")
obj.SPAGRM = SPAGRM.NullModel(ResidMatFile = ResidMatFile, 
                              SparseGRMFile = SparseGRMFile, 
                              PairwiseIBDFile = PairwiseIBDFile,
                              control = list(ControlOutlier = FALSE))

# Step 2b: perform score test
GenoFile = system.file("extdata", "simuPLINK.bed", package = "GRAB")
OutputDir = system.file("results", package = "GRAB")
OutputFile = paste0(OutputDir, "/SPAGRMMarkers.txt")
GRAB.Marker(objNull = obj.SPAGRM,
            GenoFile = GenoFile,
            OutputFile = OutputFile)
head(read.table(OutputFile, header=T))
}
