% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Simu.R
\name{GRAB.SimubVec}
\alias{GRAB.SimubVec}
\title{GRAB: simulate random effect (i.e. bVec) based on family structure}
\usage{
GRAB.SimubVec(n.fam, fam.kin, tau)
}
\arguments{
\item{n.fam}{number of families in simulation}

\item{fam.kin}{a matrix of GRM (kinship matrix)}

\item{tau}{variance component}
}
\value{
a random effect following a multivariate normal distribution
}
\description{
Simulate random effect (i.e. bVec) based on family structure
}
\examples{
fam.kin = read.table(system.file("extdata", "example_10members.kin.txt", package = "GRAB"))
fam.kin = as.matrix(fam.kin)
n.fam = 100
tau = 2
bVec = GRAB.SimubVec(n.fam, fam.kin, tau)
     
}
