% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRAB_Simu.R
\name{GRAB.SimuPheno}
\alias{GRAB.SimuPheno}
\title{Simulate phenotype using linear predictor \code{eta}}
\usage{
GRAB.SimuPheno(
  eta,
  traitType = "binary",
  control = list(pCase = 0.1, sdError = 1, pEachGroup = c(1, 1, 1), eventRate = 0.1)
)
}
\arguments{
\item{eta}{linear predictors, usually covar x beta.covar + genotype x beta.genotype}

\item{traitType}{"quantitative", "binary", "ordinal", or "time-to-event"}

\item{control}{a list of parameters for controlling the simulation process}
}
\value{
a numeric vector of phenotype
}
\description{
\code{GRAB} package can help simulate a wide variaty of phenotypes
}
\details{
Check https://wenjianbi.github.io//grab.github.io/docs/simulation_phenotype.html for more details.
}
