% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPACox.R
\name{GRAB.SPACox}
\alias{GRAB.SPACox}
\title{SPACox method in GRAB package}
\usage{
GRAB.SPACox()
}
\description{
SPACox method is to analysis time-to-event phenotype for unrelated samples in a large-scale biobank.
}
\details{
Additional list of \code{control} in \code{GRAB.NullModel()} function
Additional list of \code{control} in \code{GRAB.Marker()} function
}
\examples{
# Step 1: fit a null model
PhenoData = read.table(system.file("extdata", "example.pheno", package = "GRAB"), header = T)
obj.SPACox = GRAB.NullModel(survival::Surv(time,event) ~ Cova1 + Cova2, 
                            data = PhenoData, subjData = PhenoData$IID, 
                            method = "SPACox", traitType = "time-to-event")

# Step 2: perform score test
GenoFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.bed", package = "GRAB")
OutputDir = system.file("results", package = "GRAB")
OutputFile = paste0(OutputDir, "/SPACoxMarkers.txt")
GRAB.Marker(obj.SPACox, GenoFile = GenoFile,
            OutputFile = OutputFile)
head(read.table(OutputFile, header=T))
}
