% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGRM.R
\name{GRAB.SPAGRM}
\alias{GRAB.SPAGRM}
\title{SPAGRM method in GRAB package}
\usage{
GRAB.SPAGRM()
}
\description{
SPAGRM method is to analysis longitudinal phenotype for related samples in a large-scale biobank.
}
\details{
Please check \code{?GRAB.control} for the generic list of \code{control} in \code{GRAB.NullModel()} and \code{GRAB.Marker()}.
}
\examples{
# Step 1: fit a null model
PhenoData = read.table(system.file("extdata", "example.pheno", package = "GRAB"), header = T)
obj.SPAGRM = GRAB.NullModel(survival::Surv(time,event) ~ Cova1 + Cova2, 
                            data = PhenoData, subjData = PhenoData$IID, 
                            method = "SPAGRM", traitType = "time-to-event")

# Step 2: perform score test
GenoFile = system.file("extdata", "nSNPs-10000-nsubj-1000-ext.bed", package = "GRAB")
OutputDir = system.file("results", package = "GRAB")
OutputFile = paste0(OutputDir, "/SPAGRMMarkers.txt")
GRAB.Marker(obj.SPAGRM, GenoFile = GenoFile,
            OutputFile = OutputFile)
head(read.table(OutputFile, header=T))
}
