% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAGRM.R
\name{GRAB.SPAGRM}
\alias{GRAB.SPAGRM}
\title{SPA\if{html}{\out{<sub>}}GRM\if{html}{\out{</sub>}} method in GRAB package}
\usage{
GRAB.SPAGRM()
}
\description{
SPA\if{html}{\out{<sub>}}GRM\if{html}{\out{</sub>}} method is an empirical approach to analyzing complex traits
(including but not limited to longitudinal trait) for related samples in a
large-scale biobank. SPA\if{html}{\out{<sub>}}GRM\if{html}{\out{</sub>}} extends SPACox to support related populations.
}
\details{
See \code{\link{SPAGRM.NullModel}} for detailed instructions
on pre-calculate genotype distributions.

\strong{Additional Control Parameters for GRAB.Marker()}:
\itemize{
\item \code{zeta} (numeric, default: 0): SPA moment approximation parameter.
\item \code{tol} (numeric, default: 1e-5): Numerical tolerance for SPA convergence.
}

\strong{Marker-level results} (\code{OutputFile}) columns:
\describe{
\item{Marker}{Marker identifier (rsID or CHR:POS:REF:ALT).}
\item{Info}{Marker information in format CHR:POS:REF:ALT.}
\item{AltFreq}{Alternative allele frequency in the sample.}
\item{AltCounts}{Total count of alternative alleles.}
\item{MissingRate}{Proportion of missing genotypes.}
\item{zScore}{Z-score from the score test.}
\item{Pvalue}{P-value from the score test.}
\item{hwepval}{Hardy-Weinberg equilibrium p-value.}
}
}
\examples{
ResidMatFile <- system.file("extdata", "ResidMat.txt", package = "GRAB")
SparseGRMFile <- system.file("extdata", "SparseGRM.txt", package = "GRAB")
PairwiseIBDFile <- system.file("extdata", "PairwiseIBD.txt", package = "GRAB")
GenoFile <- system.file("extdata", "simuPLINK.bed", package = "GRAB")
OutputFile <- file.path(tempdir(), "resultSPAGRM.txt")

# Step 2a: pre-calculate genotype distributions
obj.SPAGRM <- SPAGRM.NullModel(
  ResidMatFile = ResidMatFile,
  SparseGRMFile = SparseGRMFile,
  PairwiseIBDFile = PairwiseIBDFile,
  control = list(ControlOutlier = FALSE)
)

# Step 2b: perform association tests
GRAB.Marker(obj.SPAGRM, GenoFile, OutputFile)

head(data.table::fread(OutputFile))

}
