% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAmix.R
\name{GRAB.SPAmix}
\alias{GRAB.SPAmix}
\title{SPAmix method in GRAB package}
\usage{
GRAB.SPAmix()
}
\description{
SPAmix method is to analyze complex trait (such as time-to-event data) for unrelated admixture samples.
}
\details{
Please check \code{?GRAB.control} for the generic list of \code{control} in \code{GRAB.NullModel()} and \code{GRAB.Marker()}.
}
\examples{
PhenoData = read.table(system.file("extdata", "simuPheno.txt", package = "GRAB"), header = T)
GenoFile = system.file("extdata", "simuPLINK.bed", package = "GRAB")
}
