% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WtCoxG.SPA.R
\name{GRAB.WtCoxG}
\alias{GRAB.WtCoxG}
\title{WtCoxG method in GRAB package}
\usage{
GRAB.WtCoxG(
  GenoFile,
  GenoFileIndex = NULL,
  obj.WtCoxG,
  GRM = NULL,
  OutputFile,
  control = list(AlleleOrder = "ref-first", AllMarkers = T),
  cutoff = 0.1
)
}
\arguments{
\item{GenoFile}{a character of genotype file. See Details section for more details.}

\item{GenoFileIndex}{additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{obj.WtCoxG}{a object with a class of "QCforBatchEffect"}

\item{control}{a list of parameters to decide which markers to extract.See \code{Details} section for more details}
}
\value{
an dataframe. including \code{WtCoxG.ext}, which utilizes external MAF, and \code{WtCoxG.noext} without external MAFs
}
\description{
WtCoxG method is an empirical approach to analyzing complex traits (including but not limited to time-to-event trait) by leveraging external MAFs.
}
\details{
\subsection{The details of \code{control} can be seen in \code{?GRAB.ReadGeno}}{
}
}
\examples{
setwd(system.file("WtSPAG", package = "GRAB"))
PhenoData = read.table(system.file("WtSPAG", "simuPHENO_WtSPAG.txt", package = "GRAB"), header = T)
RefPrevalence = 0.1

obj.WtCoxG = QCforBatchEffect(GenoFile = "simuBGEN1.bgen",
                             GenoFileIndex = c("simuBGEN1.bgen.bgi",
                                                "simuBGEN1.sample"),
                             OutputFile = "qcBGEN1.txt",
                             control=list(AlleleOrder = "ref-first",
                                          AllMarkers = T,
                                          IndicatorColumn = "SurvEvent", SampleIDColumn = "IID"),
                             PhenoData=PhenoData,
                             RefAfFile = "RefMAFs.txt",
                             RefPrevalence = RefPrevalence,
                             SNPnum=1e4)
names(obj.WtCoxG)

GRAB.WtCoxG(GenoFile = "simuBGEN1.bgen",
            GenoFileIndex = c("simuBGEN1.bgen.bgi", "simuBGEN1.sample"),
            obj.WtCoxG = obj.WtCoxG,
            OutputFile = "simuBGEN1.txt",
            control = list(AlleleOrder = "ref-first", AllMarkers=T))
}
