% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect.R
\name{mergeQCresults}
\alias{mergeQCresults}
\title{Merge quality control results from multiple genotype files.}
\usage{
mergeQCresults(qcResult1, ...)
}
\arguments{
\item{qcResult1}{the output R list of function QCforBatchEffect().}

\item{...}{other output R lists of function QCforBatchEffect().}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{cutoff}: a numeric. A new cutoff calculated by using merged QcResults .
\item{allcount}: a dataframe. Record the frequency of all the batch effect pvalue from \code{qcResults}
}
}
\description{
Merge quality control results from multiple genotype files.
}
