% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect.R
\name{QCforBatchEffect}
\alias{QCforBatchEffect}
\title{Quality control to check batch effect between study cohort and reference population.}
\usage{
QCforBatchEffect(
  GenoFile,
  GenoFileIndex,
  control = list(AlleleOrder = "ref-first"),
  PhenoData,
  RefAfFile,
  RefPrevalence,
  SNPnum = 10000
)
}
\arguments{
\item{GenoFile}{XXX.}

\item{GenoFileIndex}{XXX.}

\item{control}{XXX.}

\item{PhenoData}{XXX.}

\item{SNPnum}{XXX.}

\item{RefAFfile}{XXX.}

\item{RefPrev}{XXX.}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{mergeGenoInfo}: XXX.
\item{cutoff}: XXX.
\item{count}: XXX.
\item{PhenoData}: XXX.
\item{control}: XXX.
}
}
\description{
Quality control to check batch effect between study cohort and reference population.
}
\details{
something

\describe{
something
}
\subsection{The following details are about argument \code{control}}{

\describe{
Argument \code{control} includes a list of parameters for controlling the null model fitting process.
\itemize{
\item \code{maxiter}: Maximum number of iterations used to fit the null model. \emph{(default=100)}
\item \code{seed}: An integer as a random seed. Used when random process is involved. \emph{(default=12345678)}
\item \code{tolBeta}: Positive tolerance: the iterations converge when |beta - beta_old| / (|beta| + |beta_old| + tolBeta) < tolBeta. \emph{(default=0.001)}
\item \code{showInfo}: Whether to show more detailed information for trouble shooting. \emph{(default=FALSE)}
}
}
}
}
\examples{
# something of comments followed by R codes
load("/gdata01/user/liying/UKB/AD_step2/chr21_1_step2.RData")
PhenoData = PheData1[,c("ID","PC1","PC2","PC3","time","event")]\%>\%rename( SampleID =ID,Indicator=event )

qcResult1 = QCforBatchEffect(GenoFile = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c21_b0_v3.bgen")
                             ,GenoFileIndex = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb_imp_chr21_v3.bgen.bgi", 
                                                "/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c1_b0_v3_s487203.sample")
                             ,control=list(AlleleOrder = "ref-first", IDsToIncludeFile = list.files("/gdata02/master_data1/UK_Biobank/ukb22828_imp/split_R2_0.6_MAF_5e-04",paste0("chr_",21,"_chunk_",1,".txt"),full.names = T))
                             ,PhenoData=PhenoData
                             ,RefAfFile = "/gdata01/user/liying/gnomAD/uk10k.txt"
                             ,RefPrevalence = 0.02
                             ,SNPnum=1e4
)


qcResult2 = QCforBatchEffect(GenoFile = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c22_b0_v3.bgen")
                             ,GenoFileIndex = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb_imp_chr22_v3.bgen.bgi", 
                                                "/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c1_b0_v3_s487203.sample")
                             ,control=list(AlleleOrder ="ref-first", IDsToIncludeFile = list.files("/gdata02/master_data1/UK_Biobank/ukb22828_imp/split_R2_0.6_MAF_5e-04",paste0("chr_",22,"_chunk_",1,".txt"),full.names = T))
                             ,PhenoData=PhenoData
                             ,RefAfFile = "/gdata01/user/liying/gnomAD/uk10k.txt"
                             ,RefPrevalence = 0.02
                             ,SNPnum=1e4
)
}
