% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchEffect.R
\name{QCforBatchEffect}
\alias{QCforBatchEffect}
\title{Quality control to check batch effect between study cohort and reference population.}
\usage{
QCforBatchEffect(
  GenoFile,
  GenoFileIndex,
  control = list(AlleleOrder = "ref-first"),
  PhenoData,
  RefAfFile,
  RefPrevalence,
  SNPnum = 10000
)
}
\arguments{
\item{GenoFile}{a character of genotype file. See Details section for more details.}

\item{GenoFileIndex}{additional index file(s) corresponding to GenoFile. See Details section for more details.}

\item{control}{a list of parameters to decide which markers to extract.See \code{Details} section for more details}

\item{PhenoData}{a dataframe. must have at least two columns \code{SampleID} and \code{Indicator}, \code{SampleID} hold the personal identifiers for all individuals, and \code{Indicator} hold wether the event occurred, and the value must be 0 or 1 or NA}

\item{SNPnum}{a int. The least number of markers. The default is 1e4}

\item{RefAFfile}{a character of reference file. The reference file must be a \code{txt} file (header are required) including at least 7 columns \code{CHROM}, \code{POS}, \code{ID}, \code{REF}, \code{ALT}, \code{AF_ref}, \code{AN_ref}}

\item{RefPrev}{a numeric of the event rate in the population}
}
\value{
an R object with a class of "QCforBatchEffect".
\itemize{
\item{mergeGenoInfo}: a dataframe of marker info and reference MAF
\item{cutoff}: a numeric, the cut off of batcheffect.
\item{count}: a dataframe of the frequency of the batch effect pvalue.
\item{PhenoData}: a dataframe of the input PhenoData.
\item{control}: a list of parameters to decide which markers to extract.
}
}
\description{
Quality control to check batch effect between study cohort and reference population.
}
\details{
\subsection{The following details are about argument \code{control}}{

\describe{
Argument \code{control} is used to include and exclude markers. The function supports two include files of (\code{IDsToIncludeFile}, \code{RangesToIncludeFile}) and two exclude files of (\code{IDsToExcludeFile}, \code{RangesToExcludeFile}), but does not support both include and exclude files at the same time.
\itemize{
\item \code{AlleleOrder}: a character, \code{"ref-first"} or \code{"alt-first"}, to determine whether the REF/major allele should appear first or second.
\item \code{AllMarkers}:  a logical value (default: \code{FALSE}) to indicate if all markers are extracted. It might take too much memory to put genotype of all markers in R.
\item \code{IDsToIncludeFile}: a file of marker IDs to include, one column (no header). Check \code{system.file("extdata", "IDsToInclude.txt", package = "GRAB")} for an example.
\item \code{IDsToExcludeFile}: a file of marker IDs to exclude, one column (no header).
\item \code{RangesToIncludeFile}: a file of ranges to include, three columns (no headers): chromosome, start position, end position. Check \code{system.file("extdata", "RangesToInclude.txt", package = "GRAB")} for an example.
\item \code{RangesToExcludeFile}: a file of ranges to exclude, three columns (no headers): chromosome, start position, end position.
}
}
}
}
\examples{
# something of comments followed by R codes
load("/gdata01/user/liying/UKB/AD_step2/chr21_1_step2.RData")
PhenoData = PheData1[,c("ID","PC1","PC2","PC3","time","event")]\%>\%rename( SampleID =ID,Indicator=event )

qcResult1 = QCforBatchEffect(GenoFile = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c21_b0_v3.bgen")
                             ,GenoFileIndex = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb_imp_chr21_v3.bgen.bgi", 
                                                "/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c1_b0_v3_s487203.sample")
                             ,control=list(AlleleOrder = "ref-first", IDsToIncludeFile = list.files("/gdata02/master_data1/UK_Biobank/ukb22828_imp/split_R2_0.6_MAF_5e-04",paste0("chr_",21,"_chunk_",1,".txt"),full.names = T))
                             ,PhenoData=PhenoData
                             ,RefAfFile = "/gdata01/user/liying/gnomAD/uk10k.txt"
                             ,RefPrevalence = 0.02
                             ,SNPnum=1e4
)


qcResult2 = QCforBatchEffect(GenoFile = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c22_b0_v3.bgen")
                             ,GenoFileIndex = c("/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb_imp_chr22_v3.bgen.bgi", 
                                                "/gdata02/master_data1/UK_Biobank/ukb22828_imp/ukb22828_c1_b0_v3_s487203.sample")
                             ,control=list(AlleleOrder ="ref-first", IDsToIncludeFile = list.files("/gdata02/master_data1/UK_Biobank/ukb22828_imp/split_R2_0.6_MAF_5e-04",paste0("chr_",22,"_chunk_",1,".txt"),full.names = T))
                             ,PhenoData=PhenoData
                             ,RefAfFile = "/gdata01/user/liying/gnomAD/uk10k.txt"
                             ,RefPrevalence = 0.02
                             ,SNPnum=1e4
)

allQC = mergeQCresults(qcResult1,qcResult2)
Residual = getResid.wCox(formula = Surv(time, Indicator) ~ .
                         , PhenoData=PhenoData
                         , RefPrevalence=0.02)
}
